First draft revision to GOVERNANCE.md to reflect Tech Charter draft language
We recently received a proposed boilerplate Tech Charter for the Linux
Foundation which we imagine adopting in something close to its current
form.  This PR takes a pass over the GOVERNANCE.md document to try and
align it to that Tech Charter while also adding a file for the
technical charter to ultimately reside in.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>